+++ HISTORY.rst
@@ -144,6 +144,8 @@
 
 - Convert the ``MetadataLookup`` class into a general purpose template context
   registry. Add support for always available entries.
+- Add templating support to the sources inclusion and exclusion regular
+  expression fields.
 
 4.6.3 (2024-03-28)
 ==================